+++ doc/ChangeLog.md
@@ -6,6 +6,9 @@
 
 * Add support for calculations without bath.
 * Allow for read-write access to `EDIpackSolver.hloc`.
+* `EDIpackSolver.superconductive_phi` has been changed to be the modulus of the
+  computed superconductive order parameter. The complex argument of the order
+  parameter is now accessible as `EDIpackSolver.superconductive_phi_arg`.
 
 ## Version 0.7.0